[#102] Minor tidy: moved expiration's familia::base extensions to a separate file
