+++ lib/familia/features/expiration.rb
@@ -1,5 +1,7 @@
 # lib/familia/features/expiration.rb
 
+require_relative 'expiration/extensions'
+
 module Familia
   module Features
     # Expiration is a feature that provides Time To Live (TTL) management for Familia
@@ -148,6 +150,8 @@ module Familia
     #
     module Expiration
       @default_expiration = nil
+
+      Familia::Base.add_feature self, :expiration
 
       using Familia::Refinements::TimeLiterals
 
@@ -354,67 +358,6 @@ module Familia
         redis.persist(dbkey)
       end
 
-      Familia::Base.add_feature self, :expiration
-    end
-  end
-end
-
-module Familia
-  # Add a default update_expiration method for all classes that include
-  # Familia::Base. Since expiration is a core feature, we can confidently
-  # call `horreum_instance.update_expiration` without defensive programming
-  # even when expiration is not enabled for the horreum_instance class.
-  module Base
-    # Base implementation of update_expiration that maintains API compatibility
-    # with the :expiration feature's implementation.
-    #
-    # This is a no-op implementation that gets overridden by the :expiration
-    # feature. It accepts an optional default_expiration parameter to maintain
-    # interface compatibility with the overriding implementations.
-    #
-    # @param default_expiration [Numeric, nil] Time To Live in seconds
-    # @return [nil] Always returns nil for the base implementation
-    #
-    # @note This is a no-op implementation. Classes that need expiration
-    #       functionality should include the :expiration feature.
-    #
-    # @example Enable expiration feature
-    #   class MyModel < Familia::Horreum
-    #     feature :expiration
-    #     default_expiration 1.hour
-    #   end
-    #
-    def update_expiration(default_expiration: nil)
-      Familia.ld <<~LOG
-        [update_expiration] Expiration feature not enabled for #{self.class}.
-        Key: #{dbkey} Arg: #{default_expiration} (caller: #{caller(1..1)})
-      LOG
-      nil
-    end
-
-    # Base implementation of ttl that returns -1 (no expiration set)
-    #
-    # @return [Integer] Always returns -1 for the base implementation
-    #
-    def ttl
-      -1
-    end
-
-    # Base implementation of expires? that returns false
-    #
-    # @return [Boolean] Always returns false for the base implementation
-    #
-    def expires?
-      false
-    end
-
-    # Base implementation of expired? that returns false
-    #
-    # @param threshold [Numeric] Ignored in base implementation
-    # @return [Boolean] Always returns false for the base implementation
-    #
-    def expired?(_threshold = 0)
-      false
     end
   end
 end
